+++ deno/lib/external.ts
@@ -1,4 +1,3 @@
 export * from './cnx.ts';
 export * from './cvx.ts';
 export * from './ocx.ts';
-export * from './converters.ts';